Moved the first chunk of code below the chunk that loads libraries in order to make the mdy() function available.
Added code to both compliance reports that creates a column indicating whether the 60-day grace period has expired. This will be used when importing the data to Salesforce to determine whether an entity should be considered current due to a batch with a status of "UPLOADED." If the grace period hasn't expired, then an entity will be considered delinquent until the batch is processed.