Added still_missing_exp, rev, and w2 columns before the code chunk that populates their values, due to an "Uninitialized columns" warning. Added code to remove those columns after they are used to create still_missing_noticed_report.
Put the catch_during_next_cycle function inside an if statement so it won't throw an error when I run the initial compliance report.

Added a column called send_60_day_notice, which Salesforce will use upon importing the first report of a compliance cycle to determine which governments we should send 60-day notices to.

Modified the output format of the last processed and uploaded batches, so it would match the Salesforce format.

Reordered output format so it would reflect the flowpath for determining whether an LG is compliant.