Minor edits to the program description paragraph.
Revised catch_during_next_cycle() function to check for the presence of an uploaded batch since the expiration of the 60-day grace period rather than checking for a status of "Okay." The change means that the function more accurately represents our policy concerning entities that upload a batch following the 60-day grace period.

In order to accommodate the changes made to catch_during_next_cycle(), I also adjusted what I queried from Salesforce.

Finally, I removed rm(list = ls()) from the program because I learned that it is not the appropriate way to clear R.