Removed "EXP," "REV," and "PAYROLL" labels because they were unnecessary (the column name contains the information).
Changed the output format of last_processed_date and last_uploaded_date to MM/DD/YYY, in character format, to match the date format required in Salesforce (inputing information for individual entities required me to manually enter the date rather than doing a copy-and-paste).

Added and/or edited comments to the code, to better explain why or what I did.

Reformated some of the SQL queries, so they looked more organized.

Rob and I changed the format of the Report Date field in Salesforce to be in YYYY-MM-DD (character), which removed the reason I created report_date, so I removed the code that created it.

I changed one of my queries to Salesforce, so that it queried Transparency_Reporting__c rather than Transparency_Reporting__c_hd. The c_hd table closely mirrors __c, but wasn't the correct table to query.

Due to the changes I made to the output of query_last_processed_batch() and query_last_uploaded_batch(), I needed to specify which fields to convert to vectors in unnest().