I'm not sure what happened, but it looks like I have to commit my updates again (the one I did a minute ago popped back up again).
Comments here are the same as the previous commit.

I noticed that the previous commit showed that I changed valid_from_date to created_date. This is the result of changing my query to Salesforce from Transparency_Reporting__c_hd to Transparency_Reporting__c (as described in the previous commit).