Added a description to the "Output" section, describing the two additional reports created by the Compliance Report R program.
Excluded 'test city 2' from the Salesforce query.

Added code to create a report of entities that are missing three or more quarters of expense or revenue data, or more than one year of payroll data. This is meant to help us identify entities that need extra attention.

Modified the code to output the three reports.