Revised code that checks for invalid fiscal_year and type values. I added fields to both reports in order to identify the entities and batch IDs that appear in the report.
I also changed the code stating what is an allowable date, as we haven't yet decided to enforce the Transparency Board policy to have only the most recent year plus the previous four years in the non-archived data. I don't anticipate that we will enforce that soon, so manually entering the allowable fiscal_year values seems the best way to keep an eye on the fiscal_year variable.