I added LEAs and Higher Ed to this program, and had to make some changes to accommodate the differences in data sets and/or data reporting requirements.
This program doesn't consistently return a result, though it worked today. I've received error messages indicating a problem with the `unnest()` function in `generate_report()` and "R Session Aborted" messages.

The problem may be a result of all the data processing happening in the database. This program takes about four hours to run. I wonder whether the length of time it takes to complete the program and the database simultaneously responding to queries and processing new data is the source of the problem.