Added the Community Reinvestment Agency government type to the query to Salesforce.
I discovered that "Frozen" is a possible value for `Entity_Status__c`, though we rarely use it. So, I adjusted the query to Salesforce to exclude entities that are inactive and dissolved, which gets me the active entities.

I removed a query to the AWS Entity Table, as the Transparency IDs I would use from it are already reported in the data I query from Salesforce.